Changing `FLOAT` columns to `DOUBLE` columns

+++ src/main/resources/db/migration/V0.1.1__Create_contribution_methods_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE `contribution_methods` (
     `name` VARCHAR(64) NOT NULL,
     `category` ENUM('PVM', 'SKILLING', 'OTHER') NOT NULL,
     `type` ENUM('KC', 'SUBMISSION', 'XP') NOT NULL,
-    `eht_rate` FLOAT NOT NULL,
+    `eht_rate` DOUBLE NOT NULL,
     PRIMARY KEY (`id`),
     FOREIGN KEY (`tile_id`) REFERENCES `tiles` (`id`) ON DELETE CASCADE,
     CONSTRAINT name_unique_idx UNIQUE (`name`)
